Todos los services, implements, repositorys. Nuevos arreglos en dto y entidades y configuracion del model mapper

+++ src/main/java/com/proyecto/terranova/dto/CitaDTO.java
@@ -8,7 +8,7 @@ import java.time.LocalTime;
 @Data
 public class CitaDTO {
 
-    private int idCita;
+    private Long idCita;
 
     private LocalDate fechaCita;
