+++ src/main/java/com/proyecto/terranova/dto/GanadoDTO.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class GanadoDTO {
 
-    private int idGanado;
+    private Long idGanado;
 
     private String razaGanado;
